requirements.txt
# configuration
PyYAML==3.11

# thingspeak
python-dateutil==2.5.3
requests==2.11.1

# twitter
tweepy==3.5.0

# tests
pytest==3.0.1
coverage==4.1
pytest-cov==2.3.1
responses==0.5.1
testfixtures==4.10.0

# setup
bumpversion==0.5.3
